Fix compiler warning
card_eventmgr.c: In function ‘main’:
card_eventmgr.c:336:8: warning: format ‘%ld’ expects argument of type ‘long int’, but argument 5 has type ‘pid_t’ {aka ‘int’} [-Wformat=]
   DBG1("Killing process: %ld", pid);
        ^~~~~~~~~~~~~~~~~~~~~~  ~~~

+++ src/tools/card_eventmgr.c
@@ -333,7 +333,7 @@ int main(int argc, char *argv[]) {
 	    pid_t pid = read_pidfile(pidfile);
 	    if (pid > 0)
 	    {
-		DBG1("Killing process: %ld", pid);
+		DBG1("Killing process: %ld", (long)pid);
 	    	kill(pid, SIGQUIT);
 	    }
 	    else
